Move 'FeedLoaderPresentationAdaptor' to separate file
